+++ centreon-broker/storage/src/stream.cc
@@ -172,3 +172,11 @@ void stream::_update_status(std::string const& status) {
   std::lock_guard<std::mutex> lock(_statusm);
   _status = status;
 }
+
+/**
+ * @brief Method called each time cbd is reloaded.
+ */
+void stream::update() {
+  log_v2::sql()->info("storage stream update");
+  _rebuilder.force_check_rebuild_index();
+}