+++ tests/resources/Common.py
@@ -6,23 +6,42 @@ import time
 from dateutil import parser
 from datetime import datetime
 import pymysql.cursors
+from robot.libraries.BuiltIn import BuiltIn
+
 
 TIMEOUT = 30
+
+BuiltIn().import_resource('db_variables.robot')
+DB_NAME_STORAGE = BuiltIn().get_variable_value("${DBName}")
+DB_NAME_CONF = BuiltIn().get_variable_value("${DBNameConf}")
+DB_USER = BuiltIn().get_variable_value("${DBUser}")
+DB_PASS = BuiltIn().get_variable_value("${DBPass}")
+DB_HOST = BuiltIn().get_variable_value("${DBHost}")
+DB_PORT = BuiltIn().get_variable_value("${DBPort}")
 
 
 def check_connection(port: int, pid1: int, pid2: int):
     limit = time.time() + TIMEOUT
-    r = re.compile(r"^ESTAB.*127\.0\.0\.1:{}\s".format(port))
+    r = re.compile(
+        r"^ESTAB.*127\.0\.0\.1\]*:{}\s|^ESTAB.*\[::1\]*:{}\s".format(port, port))
+    p = re.compile(
+        r"127\.0\.0\.1\]*:(\d+)\s+.*127\.0\.0\.1\]*:(\d+)\s+.*,pid=(\d+)")
+    p_v6 = re.compile(
+        r"::1\]*:(\d+)\s+.*::1\]*:(\d+)\s+.*,pid=(\d+)")
     while time.time() < limit:
         out = getoutput("ss -plant")
         lst = out.split('\n')
         estab_port = list(filter(r.match, lst))
         if len(estab_port) >= 2:
             ok = [False, False]
-            p = re.compile(
-                r"127\.0\.0\.1:(\d+)\s+127\.0\.0\.1:(\d+)\s+.*,pid=(\d+)")
             for l in estab_port:
                 m = p.search(l)
+                if m is not None:
+                    if pid1 == int(m.group(3)):
+                        ok[0] = True
+                    if pid2 == int(m.group(3)):
+                        ok[1] = True
+                m = p_v6.search(l)
                 if m is not None:
                     if pid1 == int(m.group(3)):
                         ok[0] = True
@@ -31,7 +50,6 @@ def check_connection(port: int, pid1: int, pid2: int):
             if ok[0] and ok[1]:
                 return True
         time.sleep(1)
-
     return False
 
 
@@ -126,19 +144,23 @@ def create_certificate(host: str, cert: str):
     create_key_and_certificate(host, "", cert)
 
 
+def run_env():
+    return getoutput("echo $RUN_ENV | awk '{print $1}'")
+
+
 def start_mysql():
-    if not getoutput("echo $RUN_ENV | awk '{print $1}'"):
+    if not run_env():
         getoutput("systemctl start mysql")
     else:
         getoutput("mariadbd --user=root > /dev/null 2>&1 &")
 
 
 def stop_mysql():
-    if not getoutput("echo $RUN_ENV | awk '{print $1}'"):
+    if not run_env():
         getoutput("systemctl stop mysql")
     else:
         getoutput(
-        "kill -9 $(ps aux | grep 'mariadbd --user=root' | grep -v grep | awk '{print $2}')")
+            "kill -9 $(ps aux | grep 'mariadbd --user=root' | grep -v grep | awk '{print $2}')")
 
 
 def kill_broker():
@@ -183,7 +205,8 @@ def check_engine_logs_are_duplicated(log: str, date):
         idx = find_line_from(lines, date)
         count_true = 0
         count_false = 0
-        logs = []
+        logs_old = []
+        logs_new = []
         old_log = re.compile(r"\[[^\]]*\] \[[^\]]*\] ([^\[].*)")
         new_log = re.compile(
             r"\[[^\]]*\] \[[^\]]*\] \[[^\]]*\] \[[^\]]*\] (.*)")
@@ -191,25 +214,27 @@ def check_engine_logs_are_duplicated(log: str, date):
             mo = old_log.match(l)
             mn = new_log.match(l)
             if mo is not None:
-                if mo.group(1) in logs:
-                    logs.remove(mo.group(1))
+                if mo.group(1) in logs_new:
+                    logs_new.remove(mo.group(1))
                 else:
-                    logs.append(mo.group(1))
+                    logs_old.append(mo.group(1))
             else:
                 mn = new_log.match(l)
                 if mn is not None:
-                    if mn.group(1) in logs:
-                        logs.remove(mn.group(1))
+                    if mn.group(1) in logs_old:
+                        logs_old.remove(mn.group(1))
                     else:
-                        logs.append(mn.group(1))
-        if len(logs) <= 1:
+                        logs_new.append(mn.group(1))
+        if len(logs_old) <= 1:
             # It is possible to miss one log because of the initial split of the
             # file.
             return True
         else:
-            logger.console("Logs not duplicated")
-            for l in logs:
+            logger.console(
+                "{} old logs are not duplicated".format(len(logs_old)))
+            for l in logs_old:
                 logger.console(l)
+            # We don't care about new logs not duplicated, in a future, we won't have any old logs
     except IOError:
         logger.console("The file '{}' does not exist".format(log))
         return False
@@ -223,23 +248,25 @@ def find_line_from(lines, date):
     start = 0
     end = len(lines) - 1
     idx = start
-    while end - start > 1:
+    while end > start:
         idx = (start + end) // 2
         m = p.match(lines[idx])
         while m is None:
             logger.console("Unable to parse the date ({} <= {} <= {}): <<{}>>".format(
                 start, idx, end, lines[idx]))
             idx -= 1
-            if idx >= 0 :
+            if idx >= 0:
                 m = p.match(lines[idx])
             else:
                 logger.console("We are at the first line and no date found")
 
         idx_d = get_date(m.group(1))
-        if my_date <= idx_d:
+        if my_date <= idx_d and end != idx:
             end = idx
-        elif my_date > idx_d:
+        elif my_date > idx_d and start != idx:
             start = idx
+        else:
+            break
     return idx
 
 
@@ -268,6 +295,7 @@ def check_reschedule(log: str, date, content: str):
     except IOError:
         logger.console("The file '{}' does not exist".format(log))
         return False
+
 
 def check_reschedule_with_timeout(log: str, date, content: str, timeout: int):
     limit = time.time() + timeout
@@ -311,11 +339,11 @@ def set_command_status(cmd, status):
 def check_service_status_with_timeout(hostname: str, service_desc: str, status: int, timeout: int):
     limit = time.time() + timeout
     while time.time() < limit:
-        connection = pymysql.connect(host='localhost',
-                                     user='centreon',
-                                     password='centreon',
+        connection = pymysql.connect(host=DB_HOST,
+                                     user=DB_USER,
+                                     password=DB_PASS,
                                      autocommit=True,
-                                     database='centreon_storage',
+                                     database=DB_NAME_STORAGE,
                                      charset='utf8mb4',
                                      cursorclass=pymysql.cursors.DictCursor)
 
@@ -324,7 +352,7 @@ def check_service_status_with_timeout(hostname: str, service_desc: str, status: 
                 cursor.execute("SELECT s.state FROM services s LEFT JOIN hosts h ON s.host_id=h.host_id WHERE s.description=\"{}\" AND h.name=\"{}\"".format(
                     service_desc, hostname))
                 result = cursor.fetchall()
-                if result[0]['state'] and int(result[0]['state']) == status:
+                if result[0]['state'] is not None and int(result[0]['state']) == int(status):
                     return True
         time.sleep(5)
     return False
@@ -333,10 +361,10 @@ def check_service_status_with_timeout(hostname: str, service_desc: str, status: 
 def check_severity_with_timeout(name: str, level, icon_id, timeout: int):
     limit = time.time() + timeout
     while time.time() < limit:
-        connection = pymysql.connect(host='localhost',
-                                     user='centreon',
-                                     password='centreon',
-                                     database='centreon_storage',
+        connection = pymysql.connect(host=DB_HOST,
+                                     user=DB_USER,
+                                     password=DB_PASS,
+                                     database=DB_NAME_STORAGE,
                                      charset='utf8mb4',
                                      cursorclass=pymysql.cursors.DictCursor)
 
@@ -355,10 +383,10 @@ def check_severity_with_timeout(name: str, level, icon_id, timeout: int):
 def check_tag_with_timeout(name: str, typ, timeout: int):
     limit = time.time() + timeout
     while time.time() < limit:
-        connection = pymysql.connect(host='localhost',
-                                     user='centreon',
-                                     password='centreon',
-                                     database='centreon_storage',
+        connection = pymysql.connect(host=DB_HOST,
+                                     user=DB_USER,
+                                     password=DB_PASS,
+                                     database=DB_NAME_STORAGE,
                                      charset='utf8mb4',
                                      cursorclass=pymysql.cursors.DictCursor)
 
@@ -377,10 +405,10 @@ def check_tag_with_timeout(name: str, typ, timeout: int):
 def check_severities_count(value: int, timeout: int):
     limit = time.time() + timeout
     while time.time() < limit:
-        connection = pymysql.connect(host='localhost',
-                                     user='centreon',
-                                     password='centreon',
-                                     database='centreon_storage',
+        connection = pymysql.connect(host=DB_HOST,
+                                     user=DB_USER,
+                                     password=DB_PASS,
+                                     database=DB_NAME_STORAGE,
                                      charset='utf8mb4',
                                      cursorclass=pymysql.cursors.DictCursor)
 
@@ -398,10 +426,10 @@ def check_severities_count(value: int, timeout: int):
 def check_tags_count(value: int, timeout: int):
     limit = time.time() + timeout
     while time.time() < limit:
-        connection = pymysql.connect(host='localhost',
-                                     user='centreon',
-                                     password='centreon',
-                                     database='centreon_storage',
+        connection = pymysql.connect(host=DB_HOST,
+                                     user=DB_USER,
+                                     password=DB_PASS,
+                                     database=DB_NAME_STORAGE,
                                      charset='utf8mb4',
                                      cursorclass=pymysql.cursors.DictCursor)
 
@@ -419,10 +447,10 @@ def check_tags_count(value: int, timeout: int):
 def check_ba_status_with_timeout(ba_name: str, status: int, timeout: int):
     limit = time.time() + timeout
     while time.time() < limit:
-        connection = pymysql.connect(host='localhost',
-                                     user='centreon',
-                                     password='centreon',
-                                     database='centreon',
+        connection = pymysql.connect(host=DB_HOST,
+                                     user=DB_USER,
+                                     password=DB_PASS,
+                                     database=DB_NAME_CONF,
                                      charset='utf8mb4',
                                      cursorclass=pymysql.cursors.DictCursor)
         with connection:
@@ -430,7 +458,7 @@ def check_ba_status_with_timeout(ba_name: str, status: int, timeout: int):
                 cursor.execute(
                     "SELECT current_status FROM mod_bam WHERE name='{}'".format(ba_name))
                 result = cursor.fetchall()
-                if result[0]['current_status'] and int(result[0]['current_status']) == status:
+                if result[0]['current_status'] is not None and int(result[0]['current_status']) == status:
                     return True
         time.sleep(5)
     return False
@@ -439,10 +467,10 @@ def check_ba_status_with_timeout(ba_name: str, status: int, timeout: int):
 def check_service_downtime_with_timeout(hostname: str, service_desc: str, enabled, timeout: int):
     limit = time.time() + timeout
     while time.time() < limit:
-        connection = pymysql.connect(host='localhost',
-                                     user='centreon',
-                                     password='centreon',
-                                     database='centreon_storage',
+        connection = pymysql.connect(host=DB_HOST,
+                                     user=DB_USER,
+                                     password=DB_PASS,
+                                     database=DB_NAME_STORAGE,
                                      charset='utf8mb4',
                                      cursorclass=pymysql.cursors.DictCursor)
 
@@ -459,16 +487,16 @@ def check_service_downtime_with_timeout(hostname: str, service_desc: str, enable
 
 def delete_service_downtime(hst: str, svc: str):
     now = int(time.time())
-    connection = pymysql.connect(host='localhost',
-                                 user='centreon',
-                                 password='centreon',
-                                 database='centreon_storage',
+    connection = pymysql.connect(host=DB_HOST,
+                                 user=DB_USER,
+                                 password=DB_PASS,
+                                 database=DB_NAME_STORAGE,
                                  charset='utf8mb4',
                                  cursorclass=pymysql.cursors.DictCursor)
 
     with connection:
         with connection.cursor() as cursor:
-            cursor.execute("select d.internal_id from downtimes d inner join hosts h on d.host_id=h.host_id inner join services s on d.service_id=s.service_id where d.cancelled='0' and s.scheduled_downtime_depth='1' and s.description='{}' and h.name='{}'".format(svc, hst))
+            cursor.execute("select d.internal_id from downtimes d inner join hosts h on d.host_id=h.host_id inner join services s on d.service_id=s.service_id where d.cancelled='0' and s.scheduled_downtime_depth<>'0' and s.description='{}' and h.name='{}' LIMIT 1".format(svc, hst))
             result = cursor.fetchall()
             did = int(result[0]['internal_id'])
 
@@ -479,10 +507,10 @@ def delete_service_downtime(hst: str, svc: str):
 
 
 def number_of_downtimes_is(nb: int):
-    connection = pymysql.connect(host='localhost',
-                                 user='centreon',
-                                 password='centreon',
-                                 database='centreon_storage',
+    connection = pymysql.connect(host=DB_HOST,
+                                 user=DB_USER,
+                                 password=DB_PASS,
+                                 database=DB_NAME_STORAGE,
                                  charset='utf8mb4',
                                  cursorclass=pymysql.cursors.DictCursor)
 
@@ -495,10 +523,10 @@ def number_of_downtimes_is(nb: int):
 
 
 def clear_db(table: str):
-    connection = pymysql.connect(host='localhost',
-                                 user='centreon',
-                                 password='centreon',
-                                 database='centreon_storage',
+    connection = pymysql.connect(host=DB_HOST,
+                                 user=DB_USER,
+                                 password=DB_PASS,
+                                 database=DB_NAME_STORAGE,
                                  charset='utf8mb4',
                                  cursorclass=pymysql.cursors.DictCursor)
 
@@ -511,10 +539,10 @@ def clear_db(table: str):
 def check_service_severity_with_timeout(host_id: int, service_id: int, severity_id, timeout: int):
     limit = time.time() + timeout
     while time.time() < limit:
-        connection = pymysql.connect(host='localhost',
-                                     user='centreon',
-                                     password='centreon',
-                                     database='centreon_storage',
+        connection = pymysql.connect(host=DB_HOST,
+                                     user=DB_USER,
+                                     password=DB_PASS,
+                                     database=DB_NAME_STORAGE,
                                      charset='utf8mb4',
                                      cursorclass=pymysql.cursors.DictCursor)
 
@@ -532,20 +560,22 @@ def check_service_severity_with_timeout(host_id: int, service_id: int, severity_
         time.sleep(1)
     return False
 
+
 def check_host_severity_with_timeout(host_id: int, severity_id, timeout: int):
     limit = time.time() + timeout
     while time.time() < limit:
-        connection = pymysql.connect(host='localhost',
-                                     user='centreon',
-                                     password='centreon',
-                                     database='centreon_storage',
+        connection = pymysql.connect(host=DB_HOST,
+                                     user=DB_USER,
+                                     password=DB_PASS,
+                                     database=DB_NAME_STORAGE,
                                      charset='utf8mb4',
                                      autocommit=True,
                                      cursorclass=pymysql.cursors.DictCursor)
 
         with connection:
             with connection.cursor() as cursor:
-                cursor.execute("select sv.id from resources r left join severities sv ON r.severity_id=sv.severity_id where r.parent_id = 0 and r.id={}".format(host_id))
+                cursor.execute(
+                    "select sv.id from resources r left join severities sv ON r.severity_id=sv.severity_id where r.parent_id = 0 and r.id={}".format(host_id))
                 result = cursor.fetchall()
                 if len(result) > 0:
                     if severity_id == 'None':
@@ -556,7 +586,8 @@ def check_host_severity_with_timeout(host_id: int, severity_id, timeout: int):
         time.sleep(1)
     return False
 
-def check_resources_tags_with_timeout(parent_id: int, mid: int, typ: str, tag_ids: list, timeout: int):
+
+def check_resources_tags_with_timeout(parent_id: int, mid: int, typ: str, tag_ids: list, timeout: int, enabled: bool = True):
     if typ == 'servicegroup':
         t = 0
     elif typ == 'hostgroup':
@@ -567,42 +598,63 @@ def check_resources_tags_with_timeout(parent_id: int, mid: int, typ: str, tag_id
         t = 3
     limit = time.time() + timeout
     while time.time() < limit:
-        connection = pymysql.connect(host='localhost',
-                                 user='centreon',
-                                 password='centreon',
-                                 database='centreon_storage',
-                                 charset='utf8mb4',
-                                 cursorclass=pymysql.cursors.DictCursor)
+        connection = pymysql.connect(host=DB_HOST,
+                                     user=DB_USER,
+                                     password=DB_PASS,
+                                     database=DB_NAME_STORAGE,
+                                     charset='utf8mb4',
+                                     cursorclass=pymysql.cursors.DictCursor)
 
         with connection:
             with connection.cursor() as cursor:
-                cursor.execute("select t.id from resources r inner join resources_tags rt on r.resource_id=rt.resource_id inner join tags t on rt.tag_id=t.tag_id WHERE r.id={} and r.parent_id={} and t.type={}".format(mid, parent_id, t))
+                logger.console("select t.id from resources r inner join resources_tags rt on r.resource_id=rt.resource_id inner join tags t on rt.tag_id=t.tag_id WHERE r.id={} and r.parent_id={} and t.type={}".format(
+                    mid, parent_id, t))
+                cursor.execute("select t.id from resources r inner join resources_tags rt on r.resource_id=rt.resource_id inner join tags t on rt.tag_id=t.tag_id WHERE r.id={} and r.parent_id={} and t.type={}".format(
+                    mid, parent_id, t))
                 result = cursor.fetchall()
-                if len(result) > 0:
+                logger.console(result)
+                if not enabled:
+                    if len(result) == 0:
+                        return True
+                    else:
+                        for r in result:
+                            if r['id'] in tag_ids:
+                                logger.console(
+                                    "id {} is in tag ids".format(r['id']))
+                                break
+                        return True
+                elif enabled and len(result) > 0:
                     if len(result) == len(tag_ids):
                         for r in result:
                             if r['id'] not in tag_ids:
-                                logger.console("id {} is not in tag ids".format(r['id']))
+                                logger.console(
+                                    "id {} is not in tag ids".format(r['id']))
                                 break
-                            return True
+                        return True
                     else:
-                        logger.console("different sizes: result:{} and tag_ids:{}".format(len(result), len(tag_ids)))
+                        logger.console("different sizes: result:{} and tag_ids:{}".format(
+                            len(result), len(tag_ids)))
+                else:
+                    logger.console("result")
+                    logger.console(result)
         time.sleep(1)
     return False
+
 
 def check_host_tags_with_timeout(host_id: int, tag_id: int, timeout: int):
     limit = time.time() + timeout
     while time.time() < limit:
-        connection = pymysql.connect(host='localhost',
-                                 user='centreon',
-                                 password='centreon',
-                                 database='centreon_storage',
-                                 charset='utf8mb4',
-                                 cursorclass=pymysql.cursors.DictCursor)
+        connection = pymysql.connect(host=DB_HOST,
+                                     user=DB_USER,
+                                     password=DB_PASS,
+                                     database=DB_NAME_STORAGE,
+                                     charset='utf8mb4',
+                                     cursorclass=pymysql.cursors.DictCursor)
 
         with connection:
             with connection.cursor() as cursor:
-                cursor.execute("SELECT t.id FROM resources_tags rt, tags t WHERE rt.tag_id = t.tag_id and resource_id={} and t.id={}".format(host_id, tag_id))
+                cursor.execute(
+                    "SELECT t.id FROM resources_tags rt, tags t WHERE rt.tag_id = t.tag_id and resource_id={} and t.id={}".format(host_id, tag_id))
                 result = cursor.fetchall()
                 if len(result) > 0:
                     if int(result[0]['id']) == tag_id:
@@ -610,82 +662,89 @@ def check_host_tags_with_timeout(host_id: int, tag_id: int, timeout: int):
         time.sleep(1)
     return False
 
+
 def check_number_of_resources_monitored_by_poller_is(poller: int, value: int, timeout: int):
     limit = time.time() + timeout
     while time.time() < limit:
-        connection = pymysql.connect(host='localhost',
-                                 user='centreon',
-                                 password='centreon',
-                                 database='centreon_storage',
-                                 charset='utf8mb4',
-                                 cursorclass=pymysql.cursors.DictCursor)
+        connection = pymysql.connect(host=DB_HOST,
+                                     user=DB_USER,
+                                     password=DB_PASS,
+                                     database=DB_NAME_STORAGE,
+                                     charset='utf8mb4',
+                                     cursorclass=pymysql.cursors.DictCursor)
 
         with connection:
             with connection.cursor() as cursor:
-                cursor.execute("select count(*) from hosts h where instance_id={} and enabled=1".format(poller))
-                hresult = cursor.fetchall()
-                cursor.execute("select count(*) from hosts h left join services s on s.host_id=h.host_id where h.instance_id={} and s.enabled=1".format(poller))
-                sresult = cursor.fetchall()
-                if len(hresult) > 0 and len(sresult) > 0:
-                    if int(hresult[0]['count(*)']) + int(sresult[0]['count(*)']) == value:
+                cursor.execute(
+                    "SELECT count(*) FROM resources WHERE poller_id={} AND enabled=1".format(poller))
+                result = cursor.fetchall()
+                if len(result) > 0:
+                    if int(result[0]['count(*)']) == value:
                         return True
         time.sleep(1)
     return False
+
 
 def check_number_of_downtimes(expected: int, start, timeout: int):
     limit = time.time() + timeout
     d = parser.parse(start).timestamp()
     while time.time() < limit:
-        connection = pymysql.connect(host='localhost',
-                                user='centreon',
-                                password='centreon',
-                                database='centreon_storage',
-                                charset='utf8mb4',
-                                cursorclass=pymysql.cursors.DictCursor)
+        connection = pymysql.connect(host=DB_HOST,
+                                     user=DB_USER,
+                                     password=DB_PASS,
+                                     database=DB_NAME_STORAGE,
+                                     charset='utf8mb4',
+                                     cursorclass=pymysql.cursors.DictCursor)
         with connection:
             with connection.cursor() as cursor:
-                cursor.execute("SELECT count(*) FROM downtimes WHERE start_time >= {} AND deletion_time IS NULL".format(d))
+                cursor.execute(
+                    "SELECT count(*) FROM downtimes WHERE start_time >= {} AND deletion_time IS NULL".format(d))
                 result = cursor.fetchall()
                 if len(result) > 0:
-                    logger.console("{}/{} active downtimes".format(result[0]['count(*)'], expected))
+                    logger.console(
+                        "{}/{} active downtimes".format(result[0]['count(*)'], expected))
                     if int(result[0]['count(*)']) == expected:
                         return True
         time.sleep(1)
     return False
 
+
 def check_number_of_relations_between_hostgroup_and_hosts(hostgroup: int, value: int, timeout: int):
     limit = time.time() + timeout
     while time.time() < limit:
-        connection = pymysql.connect(host='localhost',
-                                 user='centreon',
-                                 password='centreon',
-                                 database='centreon_storage',
-                                 charset='utf8mb4',
-                                 cursorclass=pymysql.cursors.DictCursor)
+        connection = pymysql.connect(host=DB_HOST,
+                                     user=DB_USER,
+                                     password=DB_PASS,
+                                     database=DB_NAME_STORAGE,
+                                     charset='utf8mb4',
+                                     cursorclass=pymysql.cursors.DictCursor)
 
         with connection:
             with connection.cursor() as cursor:
-                cursor.execute("SELECT count(*) FROM hosts_hostgroups WHERE hostgroup_id={}".format(hostgroup))
+                cursor.execute(
+                    "SELECT count(*) FROM hosts_hostgroups WHERE hostgroup_id={}".format(hostgroup))
                 result = cursor.fetchall()
                 if len(result) > 0:
                     if int(result[0]['count(*)']) == value:
                         return True
         time.sleep(1)
     return False
+
 
 def check_number_of_relations_between_servicegroup_and_services(servicegroup: int, value: int, timeout: int):
     limit = time.time() + timeout
     while time.time() < limit:
-        connection = pymysql.connect(host='localhost',
-                                 user='centreon',
-                                 password='centreon',
-                                 database='centreon_storage',
-                                 charset='utf8mb4',
-                                 cursorclass=pymysql.cursors.DictCursor)
+        connection = pymysql.connect(host=DB_HOST,
+                                     user=DB_USER,
+                                     password=DB_PASS,
+                                     database=DB_NAME_STORAGE,
+                                     charset='utf8mb4',
+                                     cursorclass=pymysql.cursors.DictCursor)
 
         with connection:
             with connection.cursor() as cursor:
-                cursor.execute("SELECT count(*) FROM services_servicegroups WHERE servicegroup_id={}".format(servicegroup))
+                cursor.execute(
+                    "SELECT count(*) FROM services_servicegroups WHERE servicegroup_id={}".format(servicegroup))
                 result = cursor.fetchall()
                 if len(result) > 0:
                     if int(result[0]['count(*)']) == value:
@@ -693,26 +752,29 @@ def check_number_of_relations_between_servicegroup_and_services(servicegroup: in
         time.sleep(1)
     return False
 
+
 def check_host_status(host: str, value: int, t: int, in_resources: bool, timeout: int = TIMEOUT):
     limit = time.time() + timeout
     while time.time() < limit:
-        connection = pymysql.connect(host='localhost',
-                                 user='centreon',
-                                 password='centreon',
-                                 database='centreon_storage',
-                                 charset='utf8mb4',
-                                 cursorclass=pymysql.cursors.DictCursor)
+        connection = pymysql.connect(host=DB_HOST,
+                                     user=DB_USER,
+                                     password=DB_PASS,
+                                     database=DB_NAME_STORAGE,
+                                     charset='utf8mb4',
+                                     cursorclass=pymysql.cursors.DictCursor)
 
         with connection:
             with connection.cursor() as cursor:
                 key = ''
                 confirmed = ''
                 if in_resources:
-                    cursor.execute("SELECT status, status_confirmed FROM resources WHERE parent_id=0 AND name='{}'".format(host))
+                    cursor.execute(
+                        "SELECT status, status_confirmed FROM resources WHERE parent_id=0 AND name='{}'".format(host))
                     key = 'status'
                     confirmed = 'status_confirmed'
                 else:
-                    cursor.execute("SELECT state, state_type FROM hosts WHERE name='{}'".format(host))
+                    cursor.execute(
+                        "SELECT state, state_type FROM hosts WHERE name='{}'".format(host))
                     key = 'state'
                     confirmed = 'state_type'
                 result = cursor.fetchall()
@@ -720,6 +782,7 @@ def check_host_status(host: str, value: int, t: int, in_resources: bool, timeout
                     if int(result[0][key]) == value and int(result[0][confirmed]) == t:
                         return True
                     else:
-                        logger.console("Host '{}' has status '{}' with confirmed '{}'".format(host, result[0][key], result[0][confirmed]))
+                        logger.console("Host '{}' has status '{}' with confirmed '{}'".format(
+                            host, result[0][key], result[0][confirmed]))
         time.sleep(1)
     return False